Pull in tree from devtools
I was unable to display the Tree inside the StackingContextTree.
The nodes/tree is undefined until an action is called, and so when it goes
to render the Tree, it complains about the props being undefined. We will
have to delay the rendering of the Tree until that action is executed.